requirements.txt
apache-airflow==3.1.8
colorlog==6.10.1
pandas==3.0.1
polars==1.38.0
pyodbc==5.2.0
python-dotenv==1.2.2
Requests==2.32.5
SQLAlchemy==2.0.48
xmltodict==1.0.4
